Move tests to the right package
