+++ app/build.gradle
@@ -42,9 +42,9 @@ android {
         jvmTarget = '1.8'
     }
 
-    testOptions {
+    /*testOptions {
         unitTests.returnDefaultValues = true
-    }
+    }*/
 }
 
 dependencies {
